requirements.txt
asgiref==3.7.2
Django==4.2.1
django-autoslug==1.9.9
django-ckeditor==6.5.1
django-js-asset==2.1.0
Pillow==8.1.2
sqlparse==0.4.4
typing-extensions==4.7.1
